% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_Input_Mplus.R
\name{Import1}
\alias{Import1}
\title{An import function that will
reformat data into a usable form
from Mplus .out files. This one
is used to extract all the parameters
from the .out file when the model is a 2-class bifactor
model with 8 indicators. There is one general factor
and two specific factors.
FileName: file name of the first simulation
This has been used for multiple simulation scenarios that have
5 different sample sizes (500,1000,2000,3500,5000) and 500 nrep
per sample size. The naming convention for the different nrep
is to change the number at the end. For example:
prac_1.out
prac_2.out
prac_3.out
....
prac_500.out
This made it easier to create a loop to change the
nrep number at the end of the file name.}
\usage{
Import1(nreps, path, FileName)
}
\value{
is a usable format of the parameter loadings
}
\description{
c1: used to combine path and file name
}
